+++ storybook/webpack.config.js
@@ -1,11 +1,67 @@
+const AutoDllPlugin = require('autodll-webpack-plugin');
+
 const isCi = process.env.CI && process.env.CI !== '';
 
 module.exports = async ({ config }) => {
   const [jsxRule] = config.module.rules;
   jsxRule.use.unshift('thread-loader');
+  // Use Auto DLL plugin for faster development builds
+  if (!isCi) {
+    const [htmlWebpackPlugin] = config.plugins;
+    const { templateParameters } = htmlWebpackPlugin.options;
+    htmlWebpackPlugin.options.templateParameters = (...args) =>
+      Object.assign(templateParameters.call(null, ...args), {
+        dlls: ['./vendor.dll.js'],
+      });
+    config.plugins.push(
+      new AutoDllPlugin({
+        inject: true,
+        filename: 'vendor.dll.js',
+        entry: {
+          vendor: [
+            '@storybook/addon-actions',
+            '@storybook/addon-knobs',
+            '@storybook/addon-links',
+            '@storybook/addon-notes',
+            '@storybook/addons',
+            '@storybook/core',
+            '@storybook/react',
+            'aes-js',
+            'bignumber.js',
+            'bip39',
+            'blakejs',
+            'bs58',
+            'classnames',
+            'es6-error',
+            'faker',
+            'humanize-duration',
+            'lodash',
+            'mobx',
+            'mobx-react',
+            'mobx-react-form',
+            'mobx-react-router',
+            'moment',
+            'pbkdf2',
+            'qrcode.react',
+            'react',
+            'react-copy-to-clipboard',
+            'react-dom',
+            'react-number-format',
+            'react-router',
+            'react-svg-inline',
+            'recharts',
+            'route-parser',
+            'safe-buffer',
+            'unorm',
+            'validator',
+          ],
+        },
+      })
+    );
+  }
   return {
     ...config,
-    cache: false,
+    cache: !isCi,
     devtool: isCi ? 'none' : config.devtool,
     optimization: {
       minimize: false,